+++ models/__init__.py
@@ -1,7 +1,7 @@
-from .base import db
+from .base import db, TimestampMixin
 from .user import User
 from .project import Project
 from .comment import Comment
 from .client import Client
 
-__all__ = ["db", "User", "Project", "Comment", "Client"]
+__all__ = ["db", "TimestampMixin", "User", "Project", "Comment", "Client"]
