+++ models/client.py
@@ -1,11 +1,8 @@
-from datetime import datetime
-from .base import db
-from .user import TimestampMixin
-
+from .base import db, TimestampMixin
 
 class Client(db.Model, TimestampMixin):
-    __tablename__ = 'client'
+    __tablename__ = 'clients'
 
     id = db.Column(db.Integer, primary_key=True)
-    username = db.Column(db.String(80), nullable=False)
+    username = db.Column(db.String(80))
     email = db.Column(db.String(120), unique=True, nullable=False)
