+++ services/payment_validator.py
@@ -1,0 +1,7 @@
+class PaymentValidator:
+    """Dummy payment validator. Extend with real logic."""
+
+    @staticmethod
+    def validate(data: dict) -> bool:
+        # Placeholder for payment validation logic
+        return True
